Rearranged structure to work with new user controller class
