+++ src/validations/userValidation.js
@@ -1,13 +1,19 @@
 const inputValidation = require('./inputValidation');
 
-class SignupValidation extends inputValidation {
+class UserValidation extends inputValidation {
   constructor() {
     super();
     this.getLogger = this.getLogger.bind(this);
-    this.validateInput = this.validateInput.bind(this);
   }
 
-  validateInput(params) {
+  validateInputLogin(username, password) {
+    if (!username || !password) {
+      this.getLogger().info('Missing required parameters.');
+      throw new this.Error(403, 'Missing required parameters.');
+    }
+  }
+
+  validateInputSignup(params) {
     if (!params.username
         || !params.password
         || !params.first_name
@@ -20,4 +26,4 @@ class SignupValidation extends inputValidation {
   }
 }
 
-module.exports = SignupValidation;
+module.exports = UserValidation;
